+++ src/main/java/chikachi/discord/command/discord/CommandConfig.java
@@ -90,7 +90,7 @@ public abstract class CommandConfig {
         return false;
     }
 
-    public void read(JsonReader reader) throws IOException {
+    public final void read(JsonReader reader) throws IOException {
         JsonToken type = reader.peek();
 
         if (type == JsonToken.BOOLEAN) {
@@ -121,14 +121,18 @@ public abstract class CommandConfig {
                     }
                     reader.endArray();
                 } else {
-                    reader.skipValue();
+                    readExtra(reader, name);
                 }
             }
             reader.endObject();
         }
     }
 
-    public void write(JsonWriter writer) throws IOException {
+    protected void readExtra(JsonReader reader, String name) throws IOException {
+        reader.skipValue();
+    }
+
+    public final void write(JsonWriter writer) throws IOException {
         writer.name(this.name);
         writer.beginObject();
         writer.name("enabled");
@@ -145,6 +149,9 @@ public abstract class CommandConfig {
             writer.value(alias);
         }
         writer.endArray();
+        writeExtra(writer);
         writer.endObject();
     }
+
+    protected void writeExtra(JsonWriter writer) throws IOException {}
 }